TP 470839 - [KatnaUsdPlugins] Adding '/bigobj' flag to Windows Debug build
Summary:
It seems like at some point compile objects, while building KatanaUsdPlugins in Debug on Windowsm, were becoming too big. This fixes the problem by adding `/bigobj` flag.

This is a regression that was introduced in:
{D5968}

We might want to instead see what exactly changed and try to fix that instead.